# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My the best API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/subjects": {
            "get": {
                "tags": [
                    "subject"
                ],
                "description": "Subjects list",
                "operationId": "7ec04c0df2c6d2830d25036d6f4d4b61",
                "responses": {
                    "200": {
                        "description": "Subjects list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/Subject"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/subjects/": {
            "post": {
                "tags": [
                    "subject"
                ],
                "description": "Subject store",
                "operationId": "a36c60c53a320073a16d7060929b3bef",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Subject name",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subject create sucessfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/Subject"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/subjects/{subject}": {
            "get": {
                "tags": [
                    "subject"
                ],
                "description": "Show subject",
                "operationId": "8d258e774868d03016b16a8f4c354e04",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "Object subject",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "subject show sucessfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/Subject"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            },
            "put": {
                "tags": [
                    "subject"
                ],
                "description": "Subject Update",
                "operationId": "93b54cc3883b8b943b24c908891dbb34",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "subject name",
                        "required": true
                    },
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "Object subject",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subject update sucessfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/Subject"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            },
            "delete": {
                "tags": [
                    "subject"
                ],
                "description": "Subject delete",
                "operationId": "81e875ba9c5c29e8285393505250ffc4",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "Object subject",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subject delete sucessfull"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Subject": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "number",
                        "example": "1"
                    },
                    "name": {
                        "description": "Subject name",
                        "type": "string",
                        "example": "Subject-3212"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}